+++ kv/test_raftstore/test_test.go
@@ -702,17 +702,3 @@ func TestSplitConfChangeSnapshotUnreliableRecoverConcurrentPartition3B(t *testin
 	// Test: unreliable net, restarts, partitions, snapshots, conf change, many clients (3B) ...
 	GenericTest(t, "3B", 5, true, true, true, 100, true, true)
 }
-
-func TestDifferentDBPath(t *testing.T) {
-	cfg := config.NewTestConfig()
-	cluster := NewTestCluster(5, cfg)
-	cluster.Start()
-	defer cluster.Shutdown()
-	snapPaths := make(map[string]bool)
-	for i := 1; i <= 5; i++ {
-		path := cluster.simulator.(*NodeSimulator).nodes[uint64(i)].GetDBPath()
-		log.Infof("DBPath of Store %v: %v", i, path)
-		assert.False(t, snapPaths[path])
-		snapPaths[path] = true
-	}
-}
